requirements.txt
opencv-python
mediapipe
kafka-python
psycopg2-binary
